Adding OAth Access Token
Adding OAth access token so that we can make more than 60 requests
per minute.  I'm pushing this out to gh-pages first, and will
promptly move it back to master once testing looks good.

One issue I will need to figure out is how to make it so that this
repo can easily integrate with corporate repos. For my own personal
usecase, I'm going to be hitting paypal.github.com, so that means
I need to open an app on the PayPal flavor of github, grab that
`client_id`, and replace it here.  Its a lot of tedious work that
I don't want to give to the end user.

Hopefully I can figure something out.