UNIT TEST TIME
Allllrighty. I've only added one test, but the testing foundation
is here and ready to go!  I'm using `avajs`, partially because I
hear its awesome, partially because I want to expose myself to it
more often.

`AvaJS` uses babel to transpile the tests. I don't actually care
to transpile the DOM scripts that I am writing, so instead I am
using `babel-register` to in-place perform the transpiling step
for me.

I've also added a `setup-browser-env` file so that some of the
essential browser globals are available.