Displaying the github url before the addRepo input
One thing that bothered me about the add repo UI was that it wasn't
easy to understand what to put in the input field.  To simplify things,
I prepend `http://www.github.com/` before the input field, so that the
user knows they should be entering `:Username/:RepoName`.

If they copy-paste the entire url, it will pull out the github Url
from the input directly, taking away some of the "magic" of how the
input is handled.