Correcting past mistakes
Previously, I was maintaining "state" with a combination of an array,
 set, and map.  The array and set would use the repo's url to keep
track of the current repos, and the map made use of the id. It felt
natural to want to use unique ids returned back from the server, but
unfortunately that meant that the map would only be maintained async,
while the array and set were updated sync, causing all sorts of weird
side effects.

To correct this, I am now using the url as the key for the map, and
removed the set altogether (since map has the `has` operator available).

Now, our state is managed immediately without asynchronous side effects.

---

There are other state alternatives I could have done.  For one, I could
skip the map and just use an array, however I hate the inefficiency of
traversing through an array of objects.

I could have used a regular object literal, but order is not guaranteed
when you convert it to an array, and our `has` function gets replaced
with a truthy check, which isn't a problem, since we are expecting
either an object or `undefined`, but I kind of like the explicit
`has` check.