Converting WebWorker to SharedWorker
Converting the WebWorker into a SharedWorker.  It isn't too much
work to port over, you just need to add `port` when posting messages
out to your worker, and from the worker's side, you need to add
an `onconnect` event handle that maps the current connection to
any listeners.

Debugging with sharedworkers is kind of terrible. If something goes
wrong, the sharedworker silently fails (no stack trace, nothing).

Additionally, juggling state can feel very odd. Originally I would
load the page and pass in any/all repos via localStorage. But, with
shared workers I have persisting connections. These connections keep
track of the urls I passed in with localStorage, and when I load a
new tab I get prompted with `the (local storage url) already exists`.

So, I had to adjust the code to account for that. What I settled on
was always try to load localStorage into the sharedWorkers, but fail
silently (no notifications), and wait for the `getAllCachedRepos`
function call to fill in the blanks that I was not able to add.

At least right now, using SharedWorkers feels like a (pretty sweet)
novelty.  It is nice to see that opening multiple tabs won't clobber
Github's api, so there is that.