Revert "Converting WebWorker to SharedWorker"
This reverts commit 84d4b4d2b537d6faadbde32d012205a1d2cf27ac.

As cool as SharedWorkers are, its adding too much overhead to develop
with, and not giving me enough gains. There isn't any real good reason
for this app to support multiple tabs open at once.

So here we are...

The commit does serve as a neat demo, so I pushed the SharedWorker
commit into its own branch (`SharedWorkerExample`). And who knows,
maybe I might want to go back to it at some point.

:rocket: