Browser compatability fixes, and fixing critical bug
So I started testing with Firefox, and saw there were a few issues,
so, fixing them.

Also, yeah, noticed I made a critical bug.  After doing the github
fetch to get the token, I was mutating the `config.json` with the
access token of whoever oauth'd first...

So thats kind of bad, fixed it now though.