Making OAuth not require a page refresh
It bothered me that adding an auth thoken required a full refresh.
Adding a `window.open` is pretty trivial, and makes for a much better
experience, so I didn't have a good excuse not to.

So here we are. I've added a new route, `/auth`, that takes in the
authentication callback from github and does a postMessage back
to the original window with the authToken.

The existing `/` route does not try to make any `requests` to get
accessTokens, that is sole the responsibility of the `/auth` route.
As an added bonus, you don't end up with weird `/?code=xxx` urls everywhere.

I've also added the `origin` option that gets fed into `OctoShelf()`,
which passes in an origin for our `window.open()` call.

And finally, I am using `express-session` to store retrieved accessTokens
on `req.session`, so any/all page refreshes will come with an accessToken.