Documentation Updates, and UI overhaul
This is a lot of changes, which is why I'm tossing it up as a pull
request to visualize and digest before merging. I've changed the
homepage so that it displays the OctoShelf UI on the top, and an
informative usage guide on the bottom.

The repository bubbles dynamically update themselves based on the
window height/width. They spread out as far as possible, and if the
window is too thin or short, then the inline-view kicks in.
Additionally, I've added a button to toggle the inline view.

The more info (?) button now simply scrolls the user downward to
view more info.

---

I removed the `accessToken` config. Having the config there at all
was encouraging bad habits.  Instead, a personal_access_token can
now get passed in as an environment variable:

```
PERSONAL_ACCESS_TOKEN=xxx npm start
```