Lazy Load Background Image
I love the new background image, but because its high-res, there is
a noticable "blank screen" that happens while we wait for it to load.

To fix that, I'm adding a heavily blurred / lower quality image as the
`body` background image. I then lazy load the full-res image, and render
a `filter: blur()`'d image onto an empty-div, which I then animate
out the blur with css.

Also, added bonus, I'm using the OctoShelf icon as the Notification
icon now. :D