Fix https://github.com/TheIronDeveloper/octoshelf/issues/6
After demoing this to a few of my coworkers, a common bug that
people encountered was urls were failing due to a trailing `/`.

https://github.com/TheIronDeveloper/octoshelf
https://github.com/TheIronDeveloper/octoshelf/

are both valid repository urls, but the github api endpoint does
not seem to like the trailing slash.  So, this commit fixes that
issue by 3 ways.

* If a url with a trailing slash is pasted, the trailing slash is rm'd
* Finally, in the `addRepo` function (in the Worker), we strip trailing
  slashes.

I technically only needed to do this in the `addRepo` function, but
I liked updating the UI early (dot 1) .