Making the "About OctoShelf" section opt-in only
One thing that bothered me was baking in all the "about OctoShelf"
section into anyone that wanted to fork it.  People that fork this
webapp shouldn't have to render a usage guide.

So I am making the appDescription area partial that is rendered
by opt-in only.

To render it, you would pass in the environment variable `SHOW_APP_DESCRIPTION`

i.e

```
SHOW_APP_DESCRIPTION=true npm start
```

Otherwise, `showAppDescription` is assumed to be false.

I'm including a screenshot of the new experience (note the lack of
continued scroll bar).