Decouple the Web Worker from the rest of the app
Previously, there was an eventListener on `octoshelf.js`
 and a helper function called `parsedPostMessage`.

If the web worker posted a message, it would always go through the
eventListener found on  `octoshelf.js`, and from `octoshelf.js`,
it might call a function that was `imported` from some other module.
This fundamentally bugged me, because `octoshelf` should not be
responsible for passing functions around.

Additionally, if we wanted to `postMessage` to the web worker, I
created a helper function, `parsedPostMessage`.  Because our
modules were kind of tightly coupled, if they needed to post a message
over, they would either need to be passed the web worker itself,
or we would need to pass the `parsedPostMessage` (since it stored
a reference to the web worker.) This too, kind of sucked, since
we were making functions more tightly coupled than they should have
been.

Now, we have `Conductor`. Conductor is the traffic cordinator between
events going out the web worker, as well as events coming in.

`Conductor` handles outgoing messages by providing a high order
function that performs a postMessage. It handles inbound messages
by having a single eventListener that calls functions that have been
registered with `registerWorkerEventHandles`.

---

I will be honest, I don't know if this is really the best abstraction.

But, I like taking away unnecessary responsibility from `octoshelf.js`,
and it feels nice to not have to pass around the web worker just to
occaisionally add listeners or do postMessages.

This feels like a good approach, but I also know that by separating
things out this way, tracking the call stack of events can get confusing.
Hopefully the generous logging can help, but time will tell if this
was a good split.

For reference, logging looks like:

```
[Thing] -> [Worker] "workerFn" called with:
  params

[Worker] -> [Thing] "thingFn" called with:
  params
```