OctoWorker: Fixed bugs, added tests
So for the past few days I've noticed an awful bug.  Say you load up
Octoshelf.com, add a few repos, and leave the browser tab open for a bit.

Everything works great!  Now lets say you go home, open up your laptop,
and you end up waiting a minute to find wifi.  Within that minute,
Octoshelf will trigger a refresh, and attempt to get updated data on
the repos.  The way this was originally coded, a `removeRepo` would
get triggered for each of the active repos, removing all the repos
 from the DOM and your localStorage.

Next time you refreshed OctoShelf, it would be empty... definitely
a horrible user experience.

This PR fixes that with 2 fixes:
1. Took out the code that auto removes repos.
2. Don't make github api calls if github api is unavailble.

---

On top of that, I wrote a bunch of tests! ^_______^