Adding Navi, cleaning tests, adding `active` classes for buttons
**Adding Navi** - a small module that sends notifications.
Before the web worker use to handle sending notifications. After
toying aronud with it and trying to make updates to the notifications,
I found out that the web worker is not the best place for the notification.

For one, `window.open` is invalid within a web worker (since it has
a limited scope of document and window.) But more importantly, the
web worker's responsibility should really just be to maintain state.
Adding more work complicates the module, so... taking the notification
stuff out feels like a good call.

**Cleaning tests** - auto require babel-register and browser-setup.

**Active class** - got feedback from @ethangodt that a visual indicator
of "active" would be helpful, and hes totally right.