Adding Repo Section
Adding a Repo Section to maintain the repo `<article>`s UI. I pulled
out most of the code from `OctoShelf.js`.

`OctoShelf.js`'s responsibility is to send the githubUrl + apiUrl
to the web worker, and to load the other components.

And with this, I wrote a bunch of tests. Still have a small handful
of tests to write for `repoSection`, but I've been sitting on this
revision for a few days, so I wanted to get something out there and
think it through before green-button'ing.